src/data_pipeline/orderbook/orderbook_events_service/requirements.txt
numpy
pandas
python-binance
psycopg2